Decreased the minimum required non-ref base percentage ('biokanga align -1<snpnonrefpcnt>') from formerly 1.0 down to 0.1% for use in pooled amplicon processing workflows

+++ biokanga/Aligner.h
@@ -982,7 +982,7 @@ public:
 				int SpliceJunctLen,				// maximum splice junction length when aligning RNAseq reads
 				int MinSNPreads,				// must be at least this number of reads covering any loci before processing for SNPs at this loci
 				double QValue,					// QValue controlling FDR (Benjamini–Hochberg) SNP prediction
-				int SNPNonRefPcnt,				// only process for SNP if more/equal than this percentage number of reads are non-ref at putative SNP loci (defaults to 25) 
+				double SNPNonRefPcnt,			// only process for SNP if more/equal than this percentage number of reads are non-ref at putative SNP loci (defaults to 25.0) 
 				int MarkerLen,					// marker sequences of this length
 				double MarkerPolyThres,			// maximum allowed marker sequence base polymorphism independent of centroid SNP (default 0.333, range 0.0 to 0.5)
 				int PCRartefactWinLen,			// if >= 0 then window size to use when attempting to reduce the number of  PCR differential amplification artefacts (reads stacking to same loci)
